Extract the Global Styles Typography Panel into block-editor package
