+++ packages/block-editor/src/components/global-styles/index.js
@@ -5,3 +5,7 @@ export {
 } from './hooks';
 export { useGlobalStylesOutput } from './use-global-styles-output';
 export { GlobalStylesContext } from './context';
+export {
+	default as TypographyPanel,
+	useHasTypographyPanel,
+} from './typography-panel';
