+++ packages/edit-site/src/components/global-styles/color-utils.js
@@ -1,7 +1,6 @@
 /**
  * Internal dependencies
  */
-
 import { useSupportedStyles } from './hooks';
 
 export function useHasColorPanel( name ) {
