+++ packages/edit-site/src/components/global-styles/screen-block-list.js
@@ -10,7 +10,10 @@ import {
 } from '@wordpress/components';
 import { useSelect } from '@wordpress/data';
 import { useState, useMemo, useEffect, useRef } from '@wordpress/element';
-import { BlockIcon } from '@wordpress/block-editor';
+import {
+	BlockIcon,
+	experiments as blockEditorExperiments,
+} from '@wordpress/block-editor';
 import { useDebounce } from '@wordpress/compose';
 import { speak } from '@wordpress/a11y';
 
@@ -20,10 +23,12 @@ import { speak } from '@wordpress/a11y';
 import { useHasBorderPanel } from './border-panel';
 import { useHasColorPanel } from './color-utils';
 import { useHasDimensionsPanel } from './dimensions-panel';
-import { useHasTypographyPanel } from './typography-panel';
 import { useHasVariationsPanel } from './variations-panel';
 import ScreenHeader from './header';
 import { NavigationButtonAsItem } from './navigation-button';
+import { unlock } from '../../experiments';
+
+const { useHasTypographyPanel } = unlock( blockEditorExperiments );
 
 function useSortedBlockTypes() {
 	const blockItems = useSelect(
